Delete: Se remueve el esquema password->email. Se agrega dicho comportamiento al esquema actual.
