+++ backend_users/prod/manage.py
@@ -26,19 +26,24 @@ def seed_db():
     """Seeds db with some initial data."""
     db.session.add(UserDBModel(name="Brian",
                                lastName="Zambelli Tello",
-                               email="bzambelli@fi.uba.ar"))
+                               email="bzambelli@fi.uba.ar",
+                               password = "hola"))
     db.session.add(UserDBModel(name="Franco Martin",
                                lastName="Di Maria",
-                               email="fdimaria@fi.uba.ar"))
+                               email="fdimaria@fi.uba.ar",
+                               password = "hola"))
     db.session.add(UserDBModel(name="Hugo",
                                lastName="Larrea",
-                               email="hlarrea@fi.uba.ar"))
+                               email="hlarrea@fi.uba.ar",
+                               password = "hola"))
     db.session.add(UserDBModel(name="Juan Diego",
                                lastName="Balestieri",
-                               email="jbalestieri@fi.uba.ar"))
+                               email="jbalestieri@fi.uba.ar",
+                               password = "hola"))
     db.session.add(UserDBModel(name="Kevin",
                                lastName="Mendoza",
-                               email="kmendoza@fi.uba.ar"))
+                               email="kmendoza@fi.uba.ar",
+                               password = "hola"))
 
     db.session.commit()
